% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/no_na.R
\name{no_na}
\alias{no_na}
\title{Remove Rows with Missing Values}
\usage{
no_na(data)
}
\arguments{
\item{data}{A data frame or tibble.}
}
\value{
A data frame with all rows containing any \code{NA} values removed.
}
\description{
This function removes any rows in a data frame that contain \code{NA} values.
}
\examples{
df <- data.frame(a = c(1, NA, 3, 4), b = c("x", "y", NA, "z"))
no_na(df)
}
